Fixed deployed script
The previous version was missing files, the "files" settings in npm is
fucking confusing, so I had to include dist, bin and lib which matches
dist, dist/bin, bin, dist/lib and lib for some obscure reason.

Using dist/bin and dist/lib in this setting doesn't work.

Also, I moved babel-runtime to the dependencies because it's required in
the dist version.