Refactored create command and added tests
In an effort to reduce the amount of code in commands and therefore make
the code more testable, the logic that previously was in the commands will
be progressively moved to testable code.

As of now the create command will only make two syscalls to create the
directories, the rest of the logic lies in Migrations.

The code behind Migrations#create is more testable now.