Fixed migration table creation for postgres
When no public schema exists, the migration table creation would fail, we
now create a schema beforehand and create the migration table there.

This means that any subsequent queries must have the relation name they
refer to prefixed with the schema name.
We introduced a getTableName() method which returns the table name, this
is overriden by Pg to fix this problem.

Also, changed driver.configuration to driver.config, because it was a
terrible idea.