Add SQL preprocess function option
New config option: `preprocess`, a function that may alter the SQL code
(content of the SQL files) at runtime, just before it is executed.
The function is executed once for each SQL file of a migration.
It takes as parameters the content of the file (the SQL code) and
the relative file path, and returns the processed SQL code.

An example usage of this feature is the replacement of placeholders by
their actual values.

The default value of `preprocess` is the identity function.